% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf2incr.R
\name{acf2incr}
\alias{acf2incr}
\title{Convert position to increment autocorrelations.}
\usage{
acf2incr(acf)
}
\arguments{
\item{acf}{Length-\code{N} vector of position autocorrelations.}
}
\value{
Length \code{N-1} vector if increment autocorrelations.
}
\description{
Converts the autocorrelation of a stationary sequence \code{X} to that of its increments, \code{dX == diff(X)}.
}
\examples{
acf2incr(acf = exp(-(0:10)))
}
