% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{acf2msd}
\alias{acf2msd}
\title{Convert the Increment ACF to Position MSD}
\usage{
acf2msd(gam)
}
\arguments{
\item{gam}{the ACF at \eqn{N} lags, \eqn{ \{0, 1, ..., N-1\} }.}
}
\value{
the MSD at regular time points \eqn{ \{\Delta t, 2\Delta t, ..., N\Delta t\} }.
}
\description{
Converts the ACF of the increment of a regularly sampled stationary-increments process
\eqn{ \{X_1-X_0, X_2-X_1, ..., X_N - X_{N-1} \} } into the MSD of original process \eqn{ \{X_0, X_1, ..., X_N\} }.
}
\examples{
acf1 <- runif(10)
acf2msd(acf1)
}
