% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choleski.R
\name{Choleski}
\alias{Choleski}
\alias{cholZX}
\alias{cholXZ}
\alias{cholXZ}
\title{Toeplitz variance matrix Choleski decomposition}
\usage{
cholZX(Z, acf)

cholXZ(X, acf)
}
\arguments{
\item{Z}{\eqn{N \times P} matrix of residuals.}

\item{acf}{vector of length \eqn{N}, autocorrelation of Toeplitz matrix.}

\item{X}{\eqn{N \times P} matrix of observations.}
}
\value{
An \eqn{N \times P} matrix.
}
\description{
Compute Choleski decomposition and multiply by another matrix, or solve Choleski system of equations.
}
\details{
\code{cholZX} computes \code{X = chol(toeplitz(acf))' Z}, and \code{cholXZ} computes 
\code{Z = solve(chol(toeplitz(acf))', X)}.  Both are done with Durbin-Levinson algorithm.
}
\examples{
N <- 30
P <- 4 
Mat <- matrix(rnorm(N*P), N, P)
acf <- fbm.acf(alpha = 0.8, dT = 1/60, N = N)
cholZX(Z = Mat, acf = acf)
cholXZ(X = Mat, acf = acf)
}
