% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalToeplitz.R
\name{.DollarNames.NormalToeplitz}
\alias{.DollarNames.NormalToeplitz}
\title{# Overloaded Log-density function
#' @export
setMethod("logdens", signature(x = "NormalToeplitz", y = "ANY", z = "Toeplitz"),
          function(x, y, z) {
            .NormalToeplitz_logdens_(x$cpp_ptr, y, z$cpp_ptr)
          })}
\usage{
\method{.DollarNames}{NormalToeplitz}(x, pattern)
}
\description{
# Overloaded gradient of log-density with respect to parameters
#' @export
setMethod("grad", signature(x = "NormalToeplitz", z = "ANY", dzdt = "ANY",
                            y = "Toeplitz", dacfdt = "ANY"),
          function(x, z, dzdt, y, dacfdt) {
            .NormalToeplitz_grad_(x$cpp_ptr, z, dzdt, y$cpp_ptr, dacfdt)
          })
# Overloaded Hessian of log-density with respect to parameters
#' @export
setMethod("hess", signature(x = "NormalToeplitz", z = "ANY", dzdt = "ANY", d2zdt = "ANY",
                            y = "Toeplitz", dacfdt = "ANY", d2acfdt = "ANY"),
          function(x, z, dzdt, d2zdt, y, dacfdt, d2acfdt) {
            .NormalToeplitz_hess_(x$cpp_ptr, z, dzdt, d2zdt, y$cpp_ptr, dacfdt, d2acfdt)
          })
# Overloaded full gradient of log-density
#' @export
setMethod("grad_full", signature(x = "NormalToeplitz", z = "ANY", y = "Toeplitz"),
          function(x, z, y) {
            .NormalToeplitz_grad_full_(x$cpp_ptr, z, y$cpp_ptr)
          })
}
