% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSnorm.R
\name{dSnorm}
\alias{dSnorm}
\title{Joint density of a stationary Gaussian process}
\usage{
dSnorm(X, mu, acf, log = TRUE)
}
\arguments{
\item{X}{size \code{N x d} matrix, each column i.i.d. follows multivariate Gaussian distribution with mean \code{mu} and Toeplitz variance given by \code{acf}}

\item{mu}{length \eqn{N} vector or matrix, mean of the process}

\item{acf}{length \code{N} vector containing the first column of process variance, or a size \code{N} Toeplitz class initialized by acf}

\item{log}{logical, if \code{TRUE} returns log-densities}
}
\value{
Density (log density) of stationary Gaussian system.
}
\description{
Efficient density evaluation for the multivariate normal distribution with Toeplitz variance matrix.
}
\examples{
N <- 10
d <- 4
X <- matrix(rnorm(N*d), N, d)
theta <- 0.1
lambda <- 2

mu <- theta^2 * rep(1, N)
acf <- exp(-lambda * (1:N - 1))
acf <- Toeplitz(acf = acf)

dSnorm(X, mu, acf, log = TRUE)
}
