% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{exp1.acf}
\alias{exp1.acf}
\title{Autocorrelation of Exponential}
\usage{
exp1.acf(lambda, dT, N, incr = TRUE)
}
\arguments{
\item{lambda}{timescale.}

\item{dT}{interobservation time.}

\item{N}{Number of increment observations.}

\item{incr}{logical; whether or not to return increments.}
}
\value{
An autocorrelation vector of length \eqn{N}.
}
\description{
Autocorrelation of Exponential
}
\details{
The Exponential Autocorrelation is given by:
\deqn{\exp \{-\frac{n\Delta t}{\lambda}\}}
}
\examples{
exp.acf(lambda = 1, dT = 1/60, N = 200, incr = FALSE)
}
