% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{exp2.acf}
\alias{exp2.acf}
\title{Autocorrelation of Squared-Exponential}
\usage{
exp2.acf(lambda, dT, N, incr = TRUE)
}
\arguments{
\item{lambda}{timescale.}

\item{dT}{interobservation time.}

\item{N}{Number of increment observations.}

\item{incr}{logical; whether or not to return increments.}
}
\value{
An autocorrelation vector of length \code{N}.
}
\description{
Autocorrelation of Squared-Exponential
}
\details{
The Squared-Exponential autocorrelation is given by
}
