% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{fbm.acf}
\alias{fbm.acf}
\title{Autocorrelation of fBM Increments}
\usage{
fbm.acf(alpha, dT, N)
}
\arguments{
\item{alpha}{Subdiffusion exponent.}

\item{dT}{Interobservation time.}

\item{N}{Number of increment observations.}
}
\value{
An autocorrelation vector of length \eqn{N}.
}
\description{
Autocorrelation of fBM Increments
}
\details{
The fBM increment autocorrelation is given by:
\deqn{\frac{1}{2} \Delta t^\alpha \left[|n-1|^\alpha + |n+1|^\alpha - 2n^\alpha \right]}
}
\examples{
fbm.acf(alpha = 0.5, dT = 1/60, N = 10)
}
