% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm-msd.R
\name{fbm.msd}
\alias{fbm.msd}
\title{Mean square displacement(MSD) of fractional Brownian motion(fBM) processes}
\usage{
fbm.msd(tseq, alpha)
}
\arguments{
\item{tseq}{length \code{N} vector of timepoints}

\item{alpha}{Subdiffusion exponent of fBM process}
}
\value{
Length \code{N} vector of MSD.
}
\description{
Return the MSD of fBM at series of timepoints.
}
\details{
MSD is a measure of the average distance particles travel, defined by
\deqn{
\textrm{MSD}_X(t) = E\left[X(t) - X(0)\right]^2
}{
E[X(t) - X(0)]^2
}
The MSD of fBM is given by:
\deqn{
\textrm{MSD}(t) = t^\alpha
}{
MSD(t) = t^\alpha
}
}
\note{
In order to simulate or inference some evenly-spaced fBM processes, ACF of fBM increments can be generated by 
combining \code{fbm.msd} and \code{msd2acf}.
}
\examples{
fbm.msd(tseq = 1:10, alpha = 0.8)
}
