% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{fdyn.msd}
\alias{fdyn.msd}
\title{MSD of fBM + Dynamic Error}
\usage{
fdyn.msd(alpha, sigma, t)
}
\arguments{
\item{alpha}{Subdiffusion exponent}

\item{sigma}{Width of averaging time-window.}

\item{t}{Vector of time points}
}
\description{
MSD of fBM + Dynamic Error
}
