% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ldet}
\alias{ldet}
\title{Wrapper for the log-determinant of a matrix}
\usage{
ldet(V)
}
\arguments{
\item{V}{A square matrix.}
}
\value{
The log-determinant \code{log(det(V))}.
}
\description{
Wrapper for the log-determinant of a matrix
}
\keyword{internal}

