% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{matern.acf}
\alias{matern.acf}
\title{Autocorrelation of Matern}
\usage{
matern.acf(lambda, nu, dT, N, incr = TRUE)
}
\arguments{
\item{lambda}{timescale.}

\item{nu}{smoothness parameter.}

\item{dT}{interobservation time.}

\item{N}{Number of increment observations.}
}
\value{
An autocorrelation vector of length \code{N}.
}
\description{
Autocorrelation of Matern
}
\details{
the Matern autocorrelation is
}
