% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalToeplitz.R
\docType{class}
\name{NormalToeplitz-class}
\alias{NormalToeplitz-class}
\alias{.NormalToeplitz}
\alias{NormalToeplitz}
\title{Constructor and methods for NormalToeplitz objects.}
\usage{
NormalToeplitz(n, p)
}
\arguments{
\item{n}{Size of the Toeplitz matrix.}

\item{p}{Number of unknown parameters.}
}
\value{
A \code{NormalToeplitz} object.
}
\description{

}
\details{
It is assumed that the autocorrelation of the \code{Toeplitz} object defines a valid (i.e., positive definite) variance matrix.  The multiplication algorithms still work when this is not the case but the other algorithms do not (return values typically contain \code{NaN}s).
}

