% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalToeplitz.R
\name{NormalToeplitz}
\alias{NormalToeplitz}
\title{Multivariate normal with Toeplitz variance matrix.}
\description{
Provides methods for the Normal-Toeplitz (NTz) distribution defined as\preformatted{z ~ NTz(acf)   <=>   z ~ Normal(0, toeplitz(acf)),
}

i.e., for a multivariate normal with mean zero and variance \code{Tz = toeplitz(acf)}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NormalToeplitz$new()}}
\item \href{#method-logdens}{\code{NormalToeplitz$logdens()}}
\item \href{#method-grad}{\code{NormalToeplitz$grad()}}
\item \href{#method-hess}{\code{NormalToeplitz$hess()}}
\item \href{#method-grad_full}{\code{NormalToeplitz$grad_full()}}
\item \href{#method-clone}{\code{NormalToeplitz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$new(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Size of the Toeplitz matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{NormalToeplitz} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logdens"></a>}}
\subsection{Method \code{logdens()}}{
Log-density function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$logdens(z, acf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Density argument. A vector of length \code{n} or an \verb{n x N} matrix where each column is an observation.}

\item{\code{acf}}{A vector of length \code{n} containing the first row/column of the Toeplitz variance matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scalar or vector of length \code{N} containing the log-density of the NTz evaluated at its arguments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad"></a>}}
\subsection{Method \code{grad()}}{
Gradient of the log-density with respect to parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$grad(z, dz, acf, dacf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Density argument.  A vector of length \code{n}.}

\item{\code{dz}}{An \verb{n x ntheta} matrix containing the gradient \code{dz/dtheta}.}

\item{\code{acf}}{A vector of length \code{n} containing the first row/column of the Toeplitz variance matrix.}

\item{\code{dacf}}{An \verb{n x ntheta} matrix containing the gradient \code{dacf/dtheta}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of length \code{ntheta} containing the gradient of the NTz log-density with respect to \code{theta}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hess"></a>}}
\subsection{Method \code{hess()}}{
Hessian of log-density with respect to parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$hess(z, dz, d2z, acf, dacf, d2acf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Density argument.  A vector of length \code{n}.}

\item{\code{dz}}{An \verb{n x ntheta} matrix containing the gradient \code{dz/dtheta}.}

\item{\code{d2z}}{An \verb{n x ntheta x ntheta} array containing the Hessian \verb{d^2z/dtheta^2}.}

\item{\code{acf}}{A vector of length \code{n} containing the first row/column of the Toeplitz variance matrix.}

\item{\code{dacf}}{An \verb{n x ntheta} matrix containing the gradient \code{dacf/dtheta}.}

\item{\code{d2acf}}{An \verb{n x ntheta x ntheta} array containing the Hessian \code{dacf^2/dtheta^2}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \verb{ntheta x ntheta} matrix containing the Hessian of the NTz log-density with respect to \code{theta}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad_full"></a>}}
\subsection{Method \code{grad_full()}}{
Full gradient of log-density function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$grad_full(z, acf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{z}}{Density argument.  A vector of length \code{n}.}

\item{\code{acf}}{A vector of length \code{n} containing the first row/column of the Toeplitz variance matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \verb{n x 2} matrix containing the gradient of the NTz log-density with respect to each element of \code{z} and \code{acf}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalToeplitz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
