% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCG.R
\docType{class}
\name{PCG-class}
\alias{PCG-class}
\alias{.PCG}
\alias{PCG}
\title{Constructor and methods for PCG objects.}
\usage{
PCG(n)
}
\arguments{
\item{n}{Size of the Toeplitz matrix.}
}
\value{
A \code{PCG} object.
}
\description{
The \code{PCG} class contains efficient methods for linear algebra with symmetric positive definite (i.e., variance) Toeplitz matrices.
}
\details{
It is assumed that the autocorrelation of the \code{Toeplitz} object defines a valid (i.e., positive definite) variance matrix.  The multiplication algorithms still work when this is not the case but the other algorithms do not (return values typically contain \code{NaN}s).
}

\section{Methods}{

If \code{P1} is a \code{PCG} object, then:
\describe{
\item{\code{solve(Toep, X)}, \code{solve(Toep)}}{Solves Toeplitz systems of equations.  When second argument is missing, returns the inverse of the Toeplitz matrix.}
}
}

\examples{
# construction
N <- 5
acf <- exp(-(1:N))
P1 <- PCG(n = N)

# linear algebra
X <- rnorm(N)
P1$solve(acf = acf, y = X)
}
