% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pex-acf.R
\name{pex.acf}
\alias{pex.acf}
\title{Autocorrelation(ACF) of power exponential processes}
\usage{
pex.acf(tseq, lambda, rho)
}
\arguments{
\item{tseq}{length \code{N} vector of timepoints}

\item{lambda}{Timescale parameter}

\item{rho}{Power parameter}
}
\value{
Length \code{N} vector of power-exponential ACF.
}
\description{
Return the ACF of power exponential process at different lags
}
\details{
The power-exponential ACF is given by:
\deqn{
\textrm{acf}(t) = \exp \{-(\frac{t}{\lambda})^\rho\}
}{
\textrm{acf}(t) = exp (-(t / \lambda)^\rho)
}
}
\examples{
pex.acf(tseq = 1:10, lambda = 1, rho = 2)
}
