% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperGauss-defunct.R
\name{SuperGauss-defunct}
\alias{SuperGauss-defunct}
\alias{rSnorm}
\alias{dSnorm}
\alias{Snorm.grad}
\alias{Snorm.hess}
\title{Defunct functions in \pkg{SuperGauss}.}
\description{
Defunct functions in \pkg{SuperGauss}.
}
\section{The following functions have been removed from the \pkg{SuperGauss} package}{

\describe{
\item{\code{rSnorm()}}{Please use \code{\link[=rnormtz]{rnormtz()}} instead.}
\item{\code{dSnorm()}}{Please use \code{\link[=dnormtz]{dnormtz()}} instead.}
\item{\code{Snorm.grad()}}{Please use the \code{grad()} method in the \link{NormalToeplitz} class.}
\item{\code{Snorm.hess()}}{Please use the \code{hess()} method in the \link{NormalToeplitz} class.}
}
}

