% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperGauss.R
\docType{package}
\name{SuperGauss}
\alias{SuperGauss}
\alias{SuperGauss-package}
\title{Superfast Inference for Stationary Gaussian Distributions.}
\description{
Superfast Inference for Stationary Gaussian Distributions.
}
\details{
Add details at some point.  For now this file is used to initialize the \code{roxygen2} documentation system.
}
