% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toep-loglik.R
\name{toep.loglik}
\alias{toep.loglik}
\title{Log-likelihood function for LMN models using Superfast Algorithm}
\usage{
toep.loglik(Beta, Sigma, suff, Y, X, V, acf, Toep = Toep)
}
\description{
Log-likelihood function for LMN models using Superfast Algorithm
}

