% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toep-marg.R
\name{toep.marg}
\alias{toep.marg}
\title{Marginal Log-Posterior for the LMN Model using Superfast Algorithm}
\usage{
toep.marg(suff, Y, X, V, acf, Toep, post, prior, noSigma, debug = FALSE)
}
\arguments{
\item{suff}{The sufficient statistics to be combined with the MNIW conjugate prior.}

\item{post}{The parameters of the conditional MNIW distribution.  If missing will use \code{prior} and \code{noSigma} to calculate.}

\item{prior}{The parameters of the prior.  These are required for correct normalization.}

\item{noSigma}{Used to calculate \code{post} if it is missing.}
}
\description{
Marginal Log-Posterior for the LMN Model using Superfast Algorithm
}
