% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toep-mult.R
\name{toep.mult}
\alias{toep.mult}
\title{Toeplitz matrix multiplication}
\usage{
toep.mult(acf, X)
}
\arguments{
\item{acf}{length \code{N} vector giving the first column (or row) of the Toeplitz matrix}

\item{X}{size \code{N x d} matrix}
}
\value{
Size \code{N x d} matrix.
}
\description{
Efficient computation of symmetric Toeplitz matrix times arbitary vector or matrix using Fast Fourier Transformation.
}
\details{
Computes \code{toeplitz(acf) * X}.
}
\details{
computes \code{toeplitz(vec) * X}
computes
}
\examples{
N <- 20
d <- 3
acf <- exp(-(1:N - 1))
X <- matrix(rnorm(N*d), N, d)
toep.mult(acf, X)
}
