% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toep-post.R
\name{toep.post}
\alias{toep.post}
\title{Parameters of the Posterior Conditional MNIW Distribution}
\usage{
toep.post(suff, Y, X, V, acf, Toep, prior = NULL, noSigma = FALSE,
  calc.prior = TRUE, debug = FALSE)
}
\arguments{
\item{prior}{A list with elements \code{Lambda, Omega, Psi, nu} of the prior MNIW distribution.  Any omitted or null value defaults to zero  (see details).}

\item{noSigma}{When \code{True} assumes that \code{Sigma = diag(q)}.}

\item{calc.prior}{When \code{True} also returns the parameters of the prior distribution.}
}
\value{
A list with elements \code{Lambda, Omega, Psi, nu}, the parameters of the MNIW posterior distribution.  Also returns element \code{prior}, a list with the same elements for the conjugate prior distribution.

\code{(Beta, Sigma) | Y ~ MNIW(Lambda, Omega, Psi, nu)}.

Also returns \code{prior}, a list of the parameters of the MNIW prior.
}
\description{
Parameters of the Posterior Conditional MNIW Distribution
}
\details{
The MNIW distribution is parametrized as
\deqn{Sigma ~ iWish(Psi, nu), \qquad Beta ~ MNorm(Lambda, Omega^{-1}, Sigma.}
As this is a posterior distribution, the parameter always define a proper MNIW distribution.  \code{nu = NA} and \code{Omega = NA} respectively mean that \code{Beta = 0} or \code{Sigma = diag(q)} is known.
}
