% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toeplitz.R
\name{Toeplitz}
\alias{Toeplitz}
\alias{set_acf}
\alias{get_acf}
\alias{trace_grad}
\alias{trace_hess}
\alias{show.Toeplitz}
\alias{\%*\%}
\alias{determinant}
\alias{solve}
\alias{\%*\%,ANY,Toeplitz-method}
\alias{\%*\%,Toeplitz,ANY-method}
\alias{determinant,Toeplitz-method}
\alias{dim.Toeplitz}
\alias{ncol,Toeplitz-method}
\alias{nrow,Toeplitz-method}
\alias{show,Toeplitz-method}
\alias{solve,Toeplitz-method}
\title{Constructor and methods for Toeplitz matrix objects.}
\description{
The \code{Toeplitz} class contains efficient methods for linear algebra with symmetric positive definite (i.e., variance) Toeplitz matrices.
}
\details{
It is assumed that the autocorrelation of the \code{Toeplitz} object defines a valid (i.e., positive definite) variance matrix.  The multiplication algorithms still work when this is not the case but the other algorithms do not (return values typically contain \code{NaN}s).
}
\section{Methods}{

If \code{Tz} is a \code{Toeplitz} object with first row/column given by \code{acf}, then:
\describe{
\item{\code{Tz$set_acf(acf)}}{Sets the autocorrelation of the matrix.}
\item{\code{Tz$get_acf()}}{Gets the autocorrelation of the matrix.}
\item{\code{nrow(Tz)}, \code{ncol(Tz)}, \code{dim(Tz)}}{Selected dimension(s) of the matrix.}
\item{\code{Tz \%*\% X}, \code{X \%*\% Tz}}{Toeplitz-Matrix and Matrix-Toeplitz multiplication.  Also works if \code{X} is a vector.}
\item{\code{solve(Tz, X)}, \code{solve(Tz)}}{Solves Toeplitz systems of equations.  When second argument is missing, returns the inverse of the Toeplitz matrix.}
\item{\code{determinant(Tz)}}{Log-determinant of the Toeplitz matrix, i.e., same thing as \code{log(det(toeplitz(acf)))}.}
\item{\code{Tz$trace_grad(acf2)}}{Computes the trace of \code{solve(Tz, toeplitz(acf2))}.  This is used in the computation of the gradient of \code{log(det(Tz(theta)))} with respect to \code{theta}.}
\item{\code{Toep$trace_hess(acf2, acf3)}}{Computes the trace of\preformatted{solve(Tz, toeplitz(acf2)) \%*\% solve(Tz, toeplitz(acf3)).
}

This is used in the computation of the Hessian of \code{log(det(Tz(theta)))} with respect to \code{theta}.}
}

\subsection{\code{Tz$set_acf(acf)}}{

Sets the autocorrelation of the matrix.

}

\subsection{\code{Tz$trace_hess(acf2, acf3)}}{

Computes the trace of\preformatted{solve(Tz, toeplitz(acf2)) \%*\% solve(Tz, toeplitz(acf3)).
}

This is used in the computation of the Hessian of \code{log(det(Tz(theta)))} with respect to \code{theta}.
}
}

\examples{
# construction
acf <- exp(-(1:5))
Toep <- Toeplitz(acf = acf)
# alternatively, can allocate space first
Toep <- Toeplitz(n = length(acf))
Toep$set_acf(acf = acf)

dim(Toep) # == c(nrow(Toep), ncol(Toep))
Toep # show method
Toep$get_acf() # extract the acf

# linear algebra
X <- matrix(rnorm(10), 5, 2)
Toep \%*\% X
t(X) \%*\% Toep
solve(Toep, X)
determinant(Toep) # log-determinant
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Toeplitz$new()}}
\item \href{#method-print}{\code{Toeplitz$print()}}
\item \href{#method-size}{\code{Toeplitz$size()}}
\item \href{#method-set_acf}{\code{Toeplitz$set_acf()}}
\item \href{#method-get_acf}{\code{Toeplitz$get_acf()}}
\item \href{#method-prod}{\code{Toeplitz$prod()}}
\item \href{#method-solve}{\code{Toeplitz$solve()}}
\item \href{#method-log_det}{\code{Toeplitz$log_det()}}
\item \href{#method-trace_grad}{\code{Toeplitz$trace_grad()}}
\item \href{#method-trace_hess}{\code{Toeplitz$trace_hess()}}
\item \href{#method-clone}{\code{Toeplitz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$new(N)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_acf"></a>}}
\subsection{Method \code{set_acf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$set_acf(acf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_acf"></a>}}
\subsection{Method \code{get_acf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$get_acf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prod"></a>}}
\subsection{Method \code{prod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$prod(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-solve"></a>}}
\subsection{Method \code{solve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$solve(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_det"></a>}}
\subsection{Method \code{log_det()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$log_det()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trace_grad"></a>}}
\subsection{Method \code{trace_grad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$trace_grad(acf2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trace_hess"></a>}}
\subsection{Method \code{trace_hess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$trace_hess(acf2, acf3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Toeplitz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
