% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toeplitz.R
\docType{class}
\name{Toeplitz_Matrix-class}
\alias{Toeplitz_Matrix-class}
\alias{.Toeplitz}
\alias{setAcf}
\alias{getAcf}
\alias{traceT2}
\alias{traceT4}
\alias{show.Toeplitz_Matrix}
\alias{\%*\%}
\alias{determinant}
\alias{solve}
\alias{\%*\%,ANY,Toeplitz_Matrix-method}
\alias{\%*\%,Toeplitz_Matrix,ANY-method}
\alias{determinant,Toeplitz_Matrix-method}
\alias{dim,Toeplitz_Matrix-method}
\alias{ncol,Toeplitz_Matrix-method}
\alias{nrow,Toeplitz_Matrix-method}
\alias{show,Toeplitz_Matrix-method}
\alias{solve,Toeplitz_Matrix-method}
\alias{Toeplitz}
\title{Constructor and methods for Toeplitz matrix objects.}
\usage{
Toeplitz(n, acf)
}
\arguments{
\item{n}{size of the Toeplitz matrix.}

\item{acf}{autocorrelation vector of Toeplitz matrix.}
}
\value{
A \code{Toeplitz_Matrix} object.
}
\description{
The \code{Toeplitz_Matrix} class contains efficient methods for linear algebra with symmetric positive definite (i.e., variance) Toeplitz matrices.
}
\details{
It is assumed that the autocorrelation of the \code{Toeplitz_Matrix} defines a valid (i.e., positive definite) variance matrix.  The multiplication algorithms still work when this is not the case but the other algorithms do not (return values are typically \code{NaN}s).
}

\section{Methods}{

If \code{Toep} is a \code{Toeplitz_Matrix} object with first row/column given by \code{acf}, then:
\describe{
\item{\code{Toep$setAcf(acf)}}{Sets the autocorrelation of the matrix.}
\item{\code{Toep$getAcf()}}{Gets the autocorrelation of the matrix.}
\item{\code{nrow(Toep)}, \code{ncol(Toep)}, \code{dim(Toep)}}{Selected dimension(s) of the matrix.}
\item{\code{Toep \%*\% X}, \code{X \%*\% Toep}}{Toeplitz-Matrix and Matrix-Toeplitz multiplication.  Also works if \code{X} is a vector.}
\item{\code{solve(Toep, X)}, \code{solve(Toep)}}{Solves Toeplitz systems of equations.  When second argument is missing, returns the inverse of the Toeplitz matrix.}
\item{\code{determinant(Toep)}}{Log-determinant of the Toeplitz matrix, i.e., same thing as \code{determinant(toeplitz(acf))$modulus}.}
\item{\code{Toep$traceT2(acf2)}}{Computes the trace of \code{solve(toeplitz(acf), toeplitz(acf2))}.  This is used in the computation of the gradient of Gaussian likelihoods with Toeplitz variance matrix.}
\item{\code{Toep$traceT4(acf2, acf3)}}{Computes the trace of \code{solve(toeplitz(acf), toeplitz(acf2)) \%*\% solve(toeplitz(acf), toeplitz(acf3))}.  This is used in the computation of the Hessian of Gaussian likelihoods with Toeplitz variance matrix.}
}
}

\examples{
# construction
acf <- exp(-(1:5))
Toep <- Toeplitz(acf = acf)
\dontrun{
# alternatively, can allocate space first
Toep <- Toeplitz(n = length(acf))
Toep$setAcf(acf = acf)
}
dim(Toep)
Toep # show method
Toep$getAcf() # extract the acf

# linear algebra
X <- matrix(rnorm(10), 5, 2)
Toep \%*\% X
t(X) \%*\% Toep
solve(Toep, X)
determinant(Toep) # log-determinant
}
