% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf2incr.R
\name{acf2incr}
\alias{acf2incr}
\title{Position autocorrelation(ACF) to Increment ACF.}
\usage{
acf2incr(gam)
}
\arguments{
\item{gam}{length \code{N} vector of position ACF}
}
\value{
Length \code{N-1} vector if increment ACF.
}
\description{
Converts the ACF of a stationary sequence \code{{X0, X1, ..., XN}} to those of
its increments \code{{dX1, dX2, ..., dXN}}.
}
\examples{
acf2incr(rnorm(10))
}
