% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{acf2incr}
\alias{acf2incr}
\title{Convert the Position ACF to Increment ACF}
\usage{
acf2incr(gam)
}
\arguments{
\item{gam}{An autocorrelation sequence of length \eqn{N}.}
}
\value{
An increment autocorrelation sequence of length \eqn{N-1}.
}
\description{
Converts the autocorrelation of a stationary sequence \eqn{\{X_0, X_1, ..., X_N\}} to those of 
its increments \eqn{\{X_1-X_0, X_2-X_1, ..., X_N - X_{N-1}\} }.
}
\examples{
acf1 <- runif(10)
acf2incr(acf1)
}
