% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{acf2incr}
\alias{acf2incr}
\title{Convert the Position ACF to Increment ACF}
\usage{
acf2incr(gam)
}
\arguments{
\item{gam}{An autocorrelation sequence of length \code{nObs}.}
}
\value{
An increment autocorrelation sequence of length \code{nObs-1}.
}
\description{
Converts the autocorrelation of a stationary sequence \code{X0, X1, ..., XN} to those of its increments \code{X1-X0, X2-X1, ..., XN - X(N-1)}.
}
