% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf2incr.R
\name{acf2incr}
\alias{acf2incr}
\title{Convert position autocorrelations to increment autocorrelations.}
\usage{
acf2incr(acf)
}
\arguments{
\item{acf}{Length-\code{N} vector of position autocorrelations.}
}
\value{
Length \code{N-1} vector of increment autocorrelations.
}
\description{
Convert the autocorrelation of a stationary sequence \verb{x = (x_1, ..., x_N)} to that of its increments, \verb{dx = (x_2 - x_1, ..., x_N - x_(N-1))}.
}
\examples{
acf2incr(acf = exp(-(0:10)))
}
