% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf2msd.R
\name{acf2msd}
\alias{acf2msd}
\title{Increment autocorrelation(ACF) to Position mean square displacement(MSD).}
\usage{
acf2msd(gam)
}
\arguments{
\item{gam}{length \code{N} vector of ACF}
}
\value{
Length \code{N} vector of MSD.
}
\description{
Converts the ACF of the increment of a regularly sampled stationary-increments process
\code{{dX1, dX2, ..., dXN}} into the MSD of original process \code{{X1, ..., XN}}.
}
\examples{
acf2msd(rnorm(10))
}
