% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choleski.R
\name{Choleski}
\alias{Choleski}
\alias{cholZX}
\alias{cholXZ}
\alias{cholXZ}
\title{Toeplitz variance matrix Choleski decomposition}
\usage{
cholZX(Z, acf)

cholXZ(X, acf)
}
\arguments{
\item{Z}{\code{n x p} matrix of residuals.}

\item{acf}{vector of length \code{n}, autocorrelation of Toeplitz matrix.}

\item{X}{\code{n x p} matrix of observations.}
}
\value{
An \code{n x p} matrix.
}
\description{
Compute Choleski decomposition and multiply by another matrix, or solve Choleski system of equations.
}
\details{
\code{cholZX} computes \code{X = chol(toeplitz(acf))' Z}, and \code{cholXZ} computes \code{Z = solve(chol(toeplitz(acf))', X)}.  Both are done with Durbin-Levinson algorithm.
}
