% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choleski.R
\name{Choleski}
\alias{Choleski}
\alias{cholZX}
\alias{cholXZ}
\alias{cholXZ}
\title{Toeplitz variance matrix Choleski decomposition}
\usage{
cholZX(Z, acf)

cholXZ(X, acf)
}
\arguments{
\item{Z}{\eqn{n \times p} matrix of residuals.}

\item{acf}{vector of length \eqn{n}, autocorrelation of Toeplitz matrix.}

\item{X}{\eqn{n \times p} matrix of observations.}
}
\value{
An \eqn{n \times p} matrix.
}
\description{
Compute Choleski decomposition and multiply by another matrix, or solve Choleski system of equations.
}
\details{
\itemize{
  \item{1}{\code{cholZX} computes \code{X = chol(toeplitz(acf))' Z}}
  \item{2}{\code{cholXZ} computes\code{Z = solve(chol(toeplitz(acf))', X)}}
}
Both are done with Durbin-Levinson algorithm.
}
\examples{
n <- 30
p <- 4
Mat <- matrix(rnorm(n * p), n, p)
acf <- fbm.acf(alpha = 0.8, dT = 1/60, N = n)
cholZX(Z = Mat, acf = acf)
cholXZ(X = Mat, acf = acf)
}
