% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choleski.R
\name{Choleski}
\alias{Choleski}
\alias{cholZX}
\alias{cholXZ}
\alias{cholXZ}
\title{Toeplitz variance matrix Choleski decomposition}
\usage{
cholZX(Z, acf)

cholXZ(X, acf)
}
\arguments{
\item{Z}{size \code{N x p} matrix of residuals}

\item{acf}{length \code{N} vector, autocorrelation of Toeplitz matrix}

\item{X}{size \code{N x p} matrix of observations}
}
\value{
Size \code{N x p} matrix.
}
\description{
Compute Choleski decomposition and multiply by another matrix, or solve Choleski system of equations.
}
\details{
Both of the two functions are done with Durbin-Levinson algorithm.
\itemize{
\item \code{cholZX}: computes the Choleski decomposition of Toeplitz matrix generated by \code{acf} and multiplies arbitary matrix: X = chol(toeplitz(acf))' * Z.
\item \code{cholXZ}: computes the inversion of Choleski decomposition of Toeplitz matrix generated by \code{acf} and multiplies arbitary matrix: Z = solve(chol(toeplitz(acf)), X).
}
}
\examples{
N <- 30
p <- 4
Mat <- matrix(rnorm(N * p), N, p)
acf <- exp(-(1:N - 1))
cholZX(Z = Mat, acf = acf)
cholXZ(X = Mat, acf = acf)
}
