% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSnorm.R
\name{dSnorm}
\alias{dSnorm}
\title{Density of a Stationary Gaussian Process}
\usage{
dSnorm(X, mean, acf, log = FALSE)
}
\arguments{
\item{X}{\eqn{N \times d} matrix, d i.i.d. vector follows N(mean, Variance)}

\item{mean}{\eqn{N} vector or matrix}

\item{acf}{\eqn{N} vector or matrix, first column of variance matrix, or a Toeplitz class initialized by acf}

\item{log}{logical, if \code{TRUE} returns log-densities.}

\item{return}{vector of densities.}
}
\description{
Efficient density evaluation for the multivariate normal distribution with Toeplitz variance matrix.
}
\note{
package "SuperGauss" is required

When input {X, mean, acf} data type can be either vector or matrix, if vector, convert it into \code{n x 1} matrix
}
\examples{
N <- 30
d <- 4
X <- matrix(rnorm(N*d), N, d)
mean <- rnorm(N)
acf <- fbm.acf(alpha = 0.8, dT = 1/60, N = N)
acf <- Toeplitz(acf)
dSnorm(X, mean = mean, acf = acf, log = TRUE)
}
