% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSnorm.R
\name{dSnorm}
\alias{dSnorm}
\title{Density of a Stationary Gaussian Process}
\usage{
dSnorm(X, mean, acf, Toep, log = FALSE)
}
\arguments{
\item{X}{\code{n x d} matrix, d i.i.d. vector follows N(mean, Variance)}

\item{mean}{\code{n} vector or matrix}

\item{acf}{\code{n} vector or matrix, first column of variance matrix, or a Toeplitz class initialized by acf}

\item{log}{logical, if \code{TRUE} returns log-densities.}

\item{return}{vector of densities.}
}
\description{
Efficient density evaluation for the multivariate normal distribution with Toeplitz variance matrix.
}
\note{
package "SuperGauss" is required

When input {X, mean, acf} data type can be either vector or matrix, if vector, convert it into \code{n x 1} matrix
}

