% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSnorm.R
\name{dSnorm}
\alias{dSnorm}
\title{Density of a Stationary Gaussian Process}
\usage{
dSnorm(X, mu, acf, log = TRUE)
}
\arguments{
\item{X}{\eqn{N \times d} matrix, d i.i.d. vector follows N(mu, Variance)}

\item{mu}{\eqn{N} vector or matrix}

\item{acf}{\eqn{N} vector or matrix, first column of variance matrix, or a Toeplitz class initialized by acf}

\item{log}{logical, if \code{TRUE} returns log-densities.}

\item{return}{vector of densities.}
}
\description{
Efficient density evaluation for the multivariate normal distribution with Toeplitz variance matrix.
}
\note{
package "SuperGauss" is required

When input {X, mu, acf} data type can be either vector or matrix, if vector, convert it into \code{n x 1} matrix
}
\examples{
N <- 300
d <- 4
X <- matrix(rnorm(N*d), N, d)
theta <- 0.1
lambda <- 2

mu <- theta^2 * rep(1, N)
acf <- exp(-lambda * (1:N - 1))
acf <- Toeplitz(acf)

dSnorm(X, mu = mean, acf = acf, log = TRUE)
}
