% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{fbm.acf}
\alias{fbm.acf}
\title{Autocorrelation of fBM Increments}
\usage{
fbm.acf(alpha, dT, N)
}
\arguments{
\item{alpha}{Subdiffusion exponent.}

\item{dT}{Interobservation time.}

\item{N}{Number of increment observations.}
}
\value{
An autocorrelation vector of length \code{N}.
}
\description{
Autocorrelation of fBM Increments
}
\details{
The fBM autocorrelation is given by:
}
