% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{fdyn.msd}
\alias{fdyn.msd}
\title{MSD of fBM + Dynamic Error}
\usage{
fdyn.msd(alpha, tau, t)
}
\arguments{
\item{alpha}{Subdiffusion exponent}

\item{tau}{Width of averaging time-window.}

\item{t}{Vector of time points \eqn{ \{\Delta t, 2\Delta t, ..., N\Delta t\} }}
}
\description{
MSD of fBM + Dynamic Error
}
\details{
this function returns the MSD of \eqn{Y_t}, the integral of fBM process \eqn{X_t} with subdiffusion 
exponent \eqn{\alpha} \deqn{Y_t = \int_{0}^{\tau} X(t-s)ds}. The expression of the MSD is
\deqn{\frac{(t+\tau)^\alpha + (t-\tau)^\alpha - 2t^\alpha - 2\tau^\alpha}{(\alpha+1)(\alpha+2)}}
}
\examples{
fdyn.msd(alpha = 0.8, tau = 1/600, t = (1:200) * 1/60)
}
