% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern-acf.R
\name{matern.acf}
\alias{matern.acf}
\title{Autocorrelation(ACF) of Matern processes}
\usage{
matern.acf(tseq, lambda, nu)
}
\arguments{
\item{tseq}{length \code{N} vector of timepoints}

\item{lambda}{timescale}

\item{nu}{smoothness parameter}
}
\value{
Length \code{N} vector of matern ACF.

An ACF vector of length \code{N}.
}
\description{
Return the ACF of Matern covariance.
}
\details{
the Matern ACF is given by
\deqn{
\frac{2^{1-\nu}}{\Gamma(\nu)} \left(\sqrt{2\nu}\frac{t}{\lambda}\right)^\nu K_\nu\left(\sqrt{2\nu} \frac{t}{\lambda}\right)
}{
2^(1-v) / \Gamma(v) * (\sqrt{2v} t / \lambda)^v * K_v(\sqrt{2v} t / \lambda)
}
where K is the Bessel function of second kind.
}
\examples{
matern.acf(tseq = 1:10, lambda = 1, nu = 3/2)
}
