% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd2acf.R
\name{msd2acf}
\alias{msd2acf}
\title{Position mean square displacement(MSD) to increment autocorrelation(ACF)}
\usage{
msd2acf(eta)
}
\arguments{
\item{eta}{length \code{N} vector of MSD at regular timepoints \code{{dt, 2*dt, ..., N*dt}}}
}
\value{
Length \code{N} vector of ACF.
}
\description{
Converts the MSD of a regularly sampled stationary-increments process \code{{X0, X1, ..., XN}}
into the ACF of its increments, \code{{dX1, dX2, ..., dXN}}
}
\details{
This function only works for evenly-spaced MSD
}
\examples{
msd2acf(rnorm(10))
}
