% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{msd2acf}
\alias{msd2acf}
\title{Convert the Position MSD to Increment ACF}
\usage{
msd2acf(eta)
}
\arguments{
\item{eta}{the MSD at \eqn{N} regular time points, \eqn{ \{\Delta t, 2\Delta t, ..., N\Delta t\} }.}
}
\value{
the ACF at lags \eqn{ \{0, 1, ..., N-1\} }.
}
\description{
Converts the MSD of a regularly sampled stationary-increments process \eqn{ \{X_0, X_1, ..., X_N\} }
into the ACF of its increments, \eqn{ \{X_1-X_0, X_2-X_1, ..., X_N - X_{N-1}\} }.
}
\examples{
msd1 <- runif(10)
msd2acf(msd1)
}
