% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf-functions.R
\name{msd2acf}
\alias{msd2acf}
\title{Convert the Position MSD to Increment ACF}
\usage{
msd2acf(eta)
}
\arguments{
\item{eta}{the MSD at \code{nObs} regular time points, \code{dT, 2*dT, ..., nObs*dT}.}
}
\value{
the ACF at lags \code{0, 1, ..., nObs-1}.
}
\description{
Converts the MSD of a regularly sampled stationary-increments process \code{X0, X1, ..., XN} into the ACF of its increments, \code{X1-X0, X2-X1, ..., XN - X(N-1)}.
}
