% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSnorm.R
\name{rSnorm}
\alias{rSnorm}
\title{Simulation of a Stationary Gaussian Process.}
\usage{
rSnorm(n, acf, Z, fft = TRUE, fft.plan, ncut = 0, tol = 1e-06,
  debug = FALSE)
}
\arguments{
\item{Z}{optional \code{(2*N-2) x n} matrix of iid standard normals to use in circulant embedding.}

\item{fft}{logical, whether or not to use a superfast fft-based algorithm or the more stable Durbin-Levinson algorithm.  See details.}

\item{fft.plan}{optional plan to pass to \code{fftw}.}

\item{ncut}{length of time series to cut from the end for fft method.  This way, \code{acf} can be "padded" to embed into a circulant matrix and then cut to the correct size.  See details.}

\item{tol}{relative tolerance on negative eigenvalues.  See details.}
}
\description{
Efficient simulation of a stationary Gaussian process from a given autocorrelation function.
}
\details{
TODO: Explain exactly what ncut and tol do.  Basically, acf can be made longer than needed, and ncut values of the time series are cut off the end.  tol is used to avoid roundoff error on tiny eigenvalues which makes them look negative, i.e., all negative eigenvalues which are less than a \code{tol} fraction of the maximum eigenvalue \code{evMax} are set to \code{tol * evMax}.
}
