% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snorm-grad.R
\name{Snorm.grad}
\alias{Snorm.grad}
\title{Gradient of a stationary Gaussian log-Likelihood}
\usage{
Snorm.grad(X, mu, acf, dmu, dacf)
}
\arguments{
\item{X}{size \code{N x d} matrix, each column i.i.d. follows multivariate Gaussian distribution with mean \code{mu} and Toeplitz variance given by \code{acf}}

\item{mu}{length \code{N} vector or matrix}

\item{acf}{length \code{N} vector or matrix, first column of variance matrix, or a Toeplitz class initialized by acf}

\item{dmu}{size \code{N x p} matrix, where \code{p} is the number of parameters, each column is the partial derivative of \code{mu}}

\item{dacf}{size \code{N x p} matrix, each column is the partial derivative of \code{acf}}
}
\value{
Length \code{N} vector containing the gradient of the log-likelihood.
}
\description{
Efficient evaluation of log-likelihood gradient for stationary Gaussian data.
}
\note{
Order of partial derivative in \code{dmu} and \code{dacf} must be identical. For i-th parameter \eqn{\theta_i}, i-th
column of \code{dmu} must be \eqn{\frac{\partial \mu}{\partial \theta_i}}{d\mu/d\theta_i}
and i-th column of \code{dacf} must be 
\eqn{\frac{\partial acf}{\partial \theta_i}}{dacf/d\theta_i}
}
\examples{
N <- 300
d <- 4
X <- matrix(rnorm(N*d), N, d)
theta <- 0.1
lambda <- 2

mu <- theta^2 * rep(1, N)
acf <- exp(-lambda * (1:N - 1))
acf <- Toeplitz(acf = acf)
dmu <- matrix(0, N, 2)
dmu[, 1] <- 2 * theta * rep(1, N)
dacf <- matrix(0, N, 2)
dacf[, 2] <- -lambda * exp(-lambda * (1:N - 1))

Snorm.grad(X, mu, acf, dmu, dacf)
}
