% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snorm-hess.R
\name{Snorm.Hess}
\alias{Snorm.Hess}
\title{Hessian of a Stationary Gaussian Log-Likelihood}
\usage{
Snorm.Hess(X, mean, acf, dmean, dacf, d2mean, d2acf, Toep)
}
\arguments{
\item{X}{\code{n x d} matrix, d i.i.d. vector follows N(mean, Variance).}

\item{mean}{\code{n} vector or matrix.}

\item{acf}{\code{n} vector or matrix, first column of variance matrix}

\item{dmean}{\code{n x p} matrix, where p is the number of parameters, each column is the partial derivative of mean}

\item{dacf}{\code{n x p} matrix, each column is the partial deruvative of acf}

\item{d2mean}{\code{n x p x p} array}

\item{d2acf}{\code{n x p x p} array}

\item{Toep}{\code{n x 1} Toeplitz class, space for Toeplitz-related computation.}
}
\value{
The Hessian matrix of the log-likelihood.
}
\description{
Efficient calculation of the Hessian matrix of the log-likelihood of stationary Gaussian data.
}
\note{
package "Toeplitz" is required
}

