% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snorm-hess.R
\name{Snorm.Hess}
\alias{Snorm.Hess}
\title{Hessian of a stationary Gaussian log-Likelihood}
\usage{
Snorm.Hess(X, mu, acf, dmu, dacf, d2mu, d2acf)
}
\arguments{
\item{X}{size \code{N x d} matrix, each column i.i.d. follows multivariate Gaussian distribution with mean \code{mu} and Toeplitz variance given by \code{acf}}

\item{mu}{length \code{N} vector or matrix}

\item{acf}{length \code{N} vector or matrix, first column of variance matrix, or a Toeplitz class initialized by acf}

\item{dmu}{size \code{N x p} matrix, where \code{p} is the number of parameters, each column is the partial derivative of \code{mu}}

\item{dacf}{size \code{N x p} matrix, each column is the partial derivative of \code{acf}}

\item{d2mu}{size \code{N x p x p} array, each column is the second partial derivative of \code{mu}}

\item{d2acf}{size \code{N x p x p} array,  each column is the second partial derivative of \code{mu}}
}
\value{
Size \code{p x p} Hessian matrix of the log-likelihood.
}
\description{
Efficient calculation of the Hessian matrix of the log-likelihood of stationary Gaussian data.
}
\note{
Order of partial derivative in \code{dmu}, \code{dacf}, \code{d2mu} and \code{d2acf} must be identical. 
For i-th and j-th parameter \eqn{\theta_i, \theta_j}, ij-th
column of \code{d2mu} must be \eqn{\frac{\partial^2 \mu}{\partial \theta_i \partial \theta_j}}{d2\mu/(d\theta_i d\theta_j)}
and ij-th column of \code{d2acf} must be 
\eqn{\frac{\partial^2 acf}{\partial \theta_i \partial \theta_j}}{d2acf/(d\theta_i d\theta_j)}
}
\examples{
N <- 300
d <- 4
X <- matrix(rnorm(N*d), N, d)
theta <- 0.1
lambda <- 2

mu <- theta^2 * rep(1, N)
acf <- exp(-lambda * (1:N - 1))
acf <- Toeplitz(acf = acf)
dmu <- dacf <- matrix(0, N, 2)
dmu[, 1] <- 2 * theta * rep(1, N)
dacf[, 2] <- -lambda * exp(-lambda * (1:N - 1))
d2mu <- d2acf <- array(0, c(N, 2, 2))
d2mu[, 1, 1] <- 2 * rep(1, N)
d2acf[, 2, 2] <- lambda^2 * exp(-lambda * (1:N - 1))

Snorm.Hess(X, mu, acf, dmu, dacf, d2mu, d2acf)
}
