% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Toep-mult.R
\name{Toep.mult}
\alias{Toep.mult}
\title{Toeplitz matrix multiplies vector or matrix}
\usage{
Toep.mult(Toep, X)
}
\arguments{
\item{Toep}{vector of length \code{N} giving the first column (or row) of the Toeplitz matrix}

\item{X}{vector of length \code{N} or matrix of dimension \code{N x d}}
}
\value{
A matrix of size of X
}
\description{
efficient computation of symmetric Toeplitz matrix times arbitary vector or matrix using Fast Fourier Transformation
}
\examples{
Toep <- rnorm(200)
X <- matrix(rnorm(200 * 5), 200, 5)
Toep.mult(Toep, X)
}
